make job scheduler more modular
