+++ kitsune-fe/package.json
@@ -13,32 +13,32 @@
   },
   "dependencies": {
     "@fluent/bundle": "^0.18.0",
-    "@formkit/core": "^1.6.0",
-    "@formkit/validation": "^1.6.0",
-    "@formkit/vue": "^1.6.0",
-    "@fortawesome/fontawesome-svg-core": "^6.5.1",
-    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "@formkit/core": "^1.6.2",
+    "@formkit/validation": "^1.6.2",
+    "@formkit/vue": "^1.6.2",
+    "@fortawesome/fontawesome-svg-core": "^6.5.2",
+    "@fortawesome/free-solid-svg-icons": "^6.5.2",
     "@fortawesome/vue-fontawesome": "^3.0.6",
     "@hcaptcha/vue3-hcaptcha": "^1.3.0",
     "@mcaptcha/vanilla-glue": "^0.1.0-alpha-3",
-    "@tiptap/pm": "^2.2.4",
-    "@tiptap/starter-kit": "^2.2.4",
-    "@tiptap/vue-3": "^2.2.4",
-    "@urql/exchange-graphcache": "^6.5.0",
-    "@urql/vue": "^1.1.2",
+    "@tiptap/pm": "^2.3.0",
+    "@tiptap/starter-kit": "^2.3.0",
+    "@tiptap/vue-3": "^2.3.0",
+    "@urql/exchange-graphcache": "^7.0.0",
+    "@urql/vue": "^1.1.3",
     "@vueuse/core": "^10.9.0",
     "@zxcvbn-ts/core": "^3.0.4",
     "@zxcvbn-ts/language-common": "^3.0.4",
     "@zxcvbn-ts/language-en": "^3.0.2",
     "floating-vue": "^5.2.2",
-    "fluent-vue": "^3.5.0",
+    "fluent-vue": "^3.5.2",
     "graphql": "^16.8.1",
     "lodash": "^4.17.21",
     "pinia": "^2.1.7",
     "pinia-plugin-persistedstate": "^3.2.1",
     "rollup": "npm:@rollup/wasm-node",
     "tiptap-markdown": "^0.8.10",
-    "unhead": "^1.8.20",
+    "unhead": "^1.9.4",
     "vue": "^3.4.21",
     "vue-powerglitch": "^1.0.0",
     "vue-router": "^4.3.0",
@@ -46,26 +46,26 @@
   },
   "devDependencies": {
     "@graphql-codegen/cli": "^5.0.2",
-    "@graphql-codegen/client-preset": "^4.2.4",
+    "@graphql-codegen/client-preset": "^4.2.5",
     "@parcel/watcher": "^2.4.1",
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
     "@types/lodash": "^4.17.0",
-    "@typescript-eslint/eslint-plugin": "^7.3.1",
-    "@typescript-eslint/parser": "^7.3.1",
+    "@typescript-eslint/eslint-plugin": "^7.6.0",
+    "@typescript-eslint/parser": "^7.6.0",
     "@vitejs/plugin-vue": "^5.0.4",
     "@vue/eslint-config-prettier": "^9.0.0",
     "@vue/eslint-config-typescript": "^13.0.0",
-    "eslint": "^8.57.0",
+    "eslint": "^9.0.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-vue": "^9.23.0",
+    "eslint-plugin-vue": "^9.24.1",
     "prettier": "^3.2.5",
-    "prettier-plugin-css-order": "^2.0.1",
-    "sass": "^1.72.0",
-    "typescript": "^5.4.3",
-    "unplugin-fluent-vue": "^1.2.0",
-    "vite": "^5.2.6",
-    "vue-tsc": "^2.0.7"
+    "prettier-plugin-css-order": "^2.1.2",
+    "sass": "^1.74.1",
+    "typescript": "^5.4.4",
+    "unplugin-fluent-vue": "^1.3.0",
+    "vite": "^5.2.8",
+    "vue-tsc": "^2.0.11"
   },
   "resolutions": {
     "rollup": "npm:@rollup/wasm-node"
